backend/requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-dysmsapi==2.1.2
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
crypto==1.4.1
cryptography==45.0.3
daphne==4.1.2
Django==4.2.21
django-celery-beat==2.8.1
django-ckeditor==6.7.2
django-cors-headers==4.4.0
django-db-connection-pool==1.2.6
django-db-geventpool==4.0.8
django-filter==24.2
django-js-asset==2.2.0
django-ranged-response==0.2.0
django-redis==5.4.0
django-simple-captcha==0.6.2
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.39.1
gevent==24.2.1
gmssl==3.2.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==0.10.0
kombu==5.5.3
loguru==0.7.3
msgpack==1.1.0
Naked==0.1.32
packaging==25.0
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
py-spy==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
PyJWT==2.9.0
PyMySQL==1.0.2
pyOpenSSL==25.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
service-identity==24.2.0
shellescape==3.8.1
six==1.17.0
SQLAlchemy==2.0.41
sqlparams==6.2.0
sqlparse==0.5.3
supervisor @ git+https://github.com/Supervisor/supervisor@4bf1e57cbf292ce988dc128e0d2c8917f18da9be
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==0.18.0
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.33.0
# uWSGI==2.0.29
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.event==5.0
zope.interface==7.2
